--- web_portal/plugins/core/templates/core/includes/widgets-editor/iframe.jinja
@@ -1,12 +1,0 @@
-<section class="panel">
-    <h2>Configure</h2>
-    <form action="{{ url_for('core.post_widget_update_iframe', widget_id=dash_widget_id, back_to=back_to_url) }}"
-        method="post">
-        <label for="core-widget-iframe-src">Website Url</label>
-        <input type="url" name="src" id="core-widget-iframe-src" placeholder="https://example.com" value="{{ iframe_src }}"
-            required>
-        <label for="core-widget-iframe-height">Height (In Pixels)</label>
-        <input type="number" name="height" id="core-widget-iframe-height" min="10" value="{{ iframe_height }}" required>
-        <button type="submit">Save</button>
-    </form>
-</section>
