+++ web_portal/plugins/core/widgets.py
@@ -45,17 +45,6 @@ async def render_widget_search(config: dict) -> str:
     )
 
 
-async def render_widget_iframe(config: dict) -> str:
-    iframe_src = config.get("src")
-    iframe_height = config.get("height", 150)
-
-    return await render_template(
-        "core/includes/widgets/iframe.jinja",
-        iframe_src=iframe_src,
-        iframe_height=iframe_height,
-    )
-
-
 async def render_widget(internal_name, widget_id: int, config: dict | None) -> str:
     if config is None:
         config = {}
@@ -66,10 +55,6 @@ async def render_widget(internal_name, widget_id: int, config: dict | None) -> s
             return await render_widget_link(config.get("links", []))
         case "search":
             return await render_widget_search(config)
-        case "embed_html":
-            return config.get("content", "")
-        case "iframe":
-            return await render_widget_iframe(config)
         case _:
             logger.error(
                 "widget not found in plugin::widget_name='%s',plugin_name='core'",
@@ -111,36 +96,6 @@ async def render_widget_edit_link(
     )
 
 
-async def render_widget_edit_embed_html(
-        dash_widget_id: int,
-        config: dict | None,
-        back_to_url: str) -> str:
-    content = config.get("content", "")
-
-    return await render_template(
-        "core/includes/widgets-editor/embed_html.jinja",
-        content=content,
-        dash_widget_id=dash_widget_id,
-        back_to_url=back_to_url,
-    )
-
-
-async def render_widget_edit_iframe(
-        dash_widget_id: int,
-        config: dict | None,
-        back_to_url: str) -> str:
-    iframe_src = config.get("src", "")
-    iframe_height = config.get("height", 150)
-
-    return await render_template(
-        "core/includes/widgets-editor/iframe.jinja",
-        iframe_src=iframe_src,
-        iframe_height=iframe_height,
-        dash_widget_id=dash_widget_id,
-        back_to_url=back_to_url,
-    )
-
-
 async def render_widget_edit(
         internal_name: str,
         dash_widget_id: int,
@@ -155,10 +110,6 @@ async def render_widget_edit(
             return await render_widget_edit_link(dash_widget_id, config, back_to_url)
         case "search":
             return await render_widget_edit_search(dash_widget_id, config, back_to_url)
-        case "embed_html":
-            return await render_widget_edit_embed_html(dash_widget_id, config, back_to_url)
-        case "iframe":
-            return await render_widget_edit_iframe(dash_widget_id, config, back_to_url)
         case _:
             logger.error(
                 "widget not found in plugin::widget_name='%s',plugin_name='core'",
@@ -178,8 +129,6 @@ PLUGIN_META = PluginMeta(
         "clock": "Digital Clock",
         "links": " Links",
         "search": "Web Search",
-        "embed_html": "Embed HTML",
-        "iframe": "Embed Website",
         },
     db_models=[models],
     blueprints=[views.blueprint],
