+++ web_portal/plugins/core_extras/__init__.py
@@ -1,0 +1,2 @@
+# flake8: noqa
+from .widgets import PLUGIN_META
