+++ web_portal/plugins/core_extras/templates/core_extras/index.jinja
@@ -1,0 +1,9 @@
+{% extends "/shared/base.jinja" %}
+{% block title %}Core-Extras{% endblock %}
+{% block main %}
+<section class="panel">
+    <h2>Web Portal Core-Extras</h2>
+    <p>Welcome to Core-Extras, a built-in plugin to complement the Core plugin for Web-Portal.</p>
+</section>
+
+{% endblock %}
